backend/requirements.txt
# Core Web Frameworks
fastapi==0.104.1
uvicorn[standard]==0.24.0
pydantic==2.5.0

# AI/ML Libraries
google-generativeai==0.3.2
chromadb==0.4.22
langchain==0.1.0
langchain-google-genai==0.0.10
sentence-transformers==2.2.2

# Database & Storage
sqlalchemy==2.0.23
redis==5.0.1
pymongo==4.6.0

# Data Processing
pandas==2.1.4
numpy==1.24.3
scikit-learn==1.3.2

# NLP & Text Processing
nltk==3.8.1
spacy==3.7.2
textblob==0.17.1
langdetect==1.0.9

# Utilities
python-dotenv==1.0.0
requests==2.31.0
httpx==0.25.2
aiohttp==3.9.1
python-multipart==0.0.6

# Monitoring & Logging
psutil==5.9.6
prometheus-client==0.19.0
sentry-sdk==1.39.1

# Security
cryptography==41.0.7
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
pycryptodome==3.19.0

# Rate Limiting & Caching
slowapi==0.1.9
cachetools==5.3.2

# Development & Testing
pytest==7.4.3
pytest-asyncio==0.21.1
pytest-cov==4.1.0
black==23.12.0
flake8==6.1.0
mypy==1.7.1

# Documentation
sphinx==7.2.6
sphinx-rtd-theme==2.0.0

# Date & Time
python-dateutil==2.8.2
pytz==2023.3.post1

# File Processing
python-magic==0.4.27
PyPDF2==3.0.1
python-docx==1.1.0
openpyxl==3.1.2
markdown==3.5.1

# Optional Performance Enhancements
orjson==3.9.10
ujson==5.9.0
msgpack==1.0.7

# Environment Management
python-decouple==3.8

# CLI Tools
click==8.1.7
rich==13.7.0
typer==0.9.0

# Async Support
asyncio==3.4.3
aiofiles==23.2.1

# Vector Database Support
faiss-cpu==1.7.4
hnswlib==0.8.0

# Additional ML Tools
transformers==4.36.2
torch==2.1.2
tokenizers==0.15.0

# API Documentation
flasgger==0.9.7.1
pydantic-settings==2.1.0
